add a styles folder and all the styles in the comps and screens move from the comps to the folder
