+++ src/components/UserModal/UserModal.js
@@ -14,8 +14,8 @@ import { GENDERS } from "../../constants/Genders";
 import { PREVIUS_SCREEN } from "../../constants/Preivus_screens";
 import { MODAL } from "../../constants/Modal";
 import useModal from "../../customHooks/UserModal";
-const WIDTH = Dimensions.get("screen").width;
-const HEIGHT = Dimensions.get("screen").height;
+import { userModalStyles } from "../../styles/userModalStyles";
+
 const UserModal = ({ user, setShowModal, previusScreen }) => {
   const {
     firstName,
@@ -45,16 +45,16 @@ const UserModal = ({ user, setShowModal, previusScreen }) => {
   const [openList, setOpenList] = useState(false);
 
   return (
-    <TouchableOpacity disabled={true} style={styles.container}>
-      <View style={styles.modal}>
-        <View style={styles.header}>
+    <TouchableOpacity disabled={true} style={userModalStyles.container}>
+      <View style={userModalStyles.modal}>
+        <View style={userModalStyles.header}>
           <Text_uk
             textValue={
               previusScreen === PREVIUS_SCREEN.EDIT ? MODAL.EDIT : MODAL.ADD
             }
             textValue2={MODAL.USER}
           />
-          <View style={styles.exit}>
+          <View style={userModalStyles.exit}>
             <Ionicons
               name="exit-sharp"
               size={24}
@@ -63,7 +63,7 @@ const UserModal = ({ user, setShowModal, previusScreen }) => {
             />
           </View>
         </View>
-        <View style={styles.content}>
+        <View style={userModalStyles.content}>
           <ScrollView>
             <DropDownList_uk
               data={genders}
@@ -110,51 +110,12 @@ const UserModal = ({ user, setShowModal, previusScreen }) => {
             />
           </ScrollView>
         </View>
-        <View style={styles.saveBtn}>
+        <View style={userModalStyles.saveBtn}>
           <CustomBtn_uk handleFunc={handleSave} textValue={MODAL.SAVE} />
         </View>
       </View>
     </TouchableOpacity>
   );
 };
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  modal: {
-    height: HEIGHT - 200,
-    width: WIDTH - 80,
-    backgroundColor: "#3a4f50",
-    borderRadius: 10,
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  content: {
-    flex: 1,
-    padding: 20,
-  },
-  input: {
-    padding: 10,
-    borderRadius: 10,
-    backgroundColor: "#1b2223",
-    marginVertical: 20,
-  },
-  header: {
-    alignItems: "center",
-    flexDirection: "row",
-    gap: 70,
-    padding: 10,
-  },
-  exit: {
-    alignSelf: "flex-end",
-  },
-  saveBtn: {
-    backgroundColor: "darkblue",
-    padding: 7,
-    marginBottom: 20,
-    borderRadius: 50,
-  },
-});
+
 export default UserModal;
