+++ src/constants/Header.js
@@ -1,4 +1,6 @@
 export const HEADER = {
-     ADMIN_PANEL: "Admin Panel😎",
-     ADD_USER:"Add user"
+  ADMIN_PANEL: "Admin Panel😎",
+  ADD_USER: "Add user",
+  SEARCH_BY_EMAIL: "Search By Email..",
+  SEARCH_BY_NAME: "Search By Name..",
 };
