+++ src/styles/adminPanelStyles.js
@@ -1,0 +1,7 @@
+import { StyleSheet } from "react-native";
+
+export const adminPanelStyles = StyleSheet.create({
+     container: {
+       flex: 8,
+     },
+   });
