+++ src/styles/deleteAlertStyles.js
@@ -1,0 +1,13 @@
+import { StyleSheet } from "react-native";
+
+export const deleteAlertStyles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: "center",
+  },
+  content: {
+    flexDirection: "row",
+    justifyContent: "space-between",
+    gap: 100,
+  },
+});
